Module uosql::net

The network api

This module defines types and functions to read and write objects from and to a TcpStream. However, the types in this module are types that are passed to other methods and thus more "high level". They do not represent the byte layout of objects in the protocol!

Protocol

All communication is send using TCP, which emulates a data stream. On top of TCP, this database sends single packets.

Every packet begins with a four byte length field that contains the size of the packet in network byte order.

...

Structs

Greeting
This is the first packet being sent by the server after the TCP connection is established.

Login
The client responds with this packet to a Greeting packet, finishing the authentication handshake.

Response
Sent by the server to the client.

Enums

Command
Sent by the client to the server.