Struct uosql::storage::Table

pub struct Table;

A database table

Through this type, you can retreive certain meta information about the table (like column names, column types, storage engine, ...). It's access method locks the table globally and returns a storage engine to access the table data.

Methods

impl Table

fn load(database: &str, table: &str) -> Table

fn columns(&self) -> &[Column]